move the tag initializer logic to an individual file
